First usage of segment reporting
