+++ Makefile
@@ -30,7 +30,7 @@ PKG := github.com/drud/ddev
 SRC_DIRS := cmd pkg
 
 # Version variables to replace in build
-VERSION_VARIABLES ?= DdevVersion SentryDSN
+VERSION_VARIABLES ?= DdevVersion SentryDSN SegmentKey
 
 # These variables will be used as the default unless overridden by the make
 DdevVersion ?= $(VERSION)
