+++ cmd/ddev/cmd/root.go
@@ -8,7 +8,6 @@ import (
 	"github.com/drud/ddev/pkg/updatecheck"
 	"github.com/drud/ddev/pkg/util"
 	"github.com/drud/ddev/pkg/version"
-	"github.com/getsentry/raven-go"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -95,33 +94,6 @@ Support: https://ddev.readthedocs.io/en/stable/#support`,
 		}
 
 	},
-	PersistentPostRun: func(cmd *cobra.Command, args []string) {
-		// Do not report these comamnds
-		ignores := map[string]bool{"list": true, "version": true, "help": true, "auth-pantheon": true, "hostname": true}
-		if _, ok := ignores[cmd.CalledAs()]; ok {
-			return
-		}
-		sentryNotSetupWarning()
-
-		// All this nonsense is to capture the official usage we used for this command.
-		// Unfortunately cobra doesn't seem to provide this easily.
-		// We use the first word of Use: to get it.
-		cmdCopy := cmd
-		var fullCommand = make([]string, 0)
-		fullCommand = append(fullCommand, util.GetFirstWord(cmdCopy.Use))
-		for cmdCopy.HasParent() {
-			fullCommand = append(fullCommand, util.GetFirstWord(cmdCopy.Parent().Use))
-			cmdCopy = cmdCopy.Parent()
-		}
-		uString := "Usage:"
-		for i := len(fullCommand) - 1; i >= 0; i = i - 1 {
-			uString = uString + " " + fullCommand[i]
-		}
-
-		if globalconfig.DdevGlobalConfig.InstrumentationOptIn && version.SentryDSN != "" && nodeps.IsInternetActive() {
-			_ = raven.CaptureMessageAndWait(uString, map[string]string{"severity-level": "info", "report-type": "usage"})
-		}
-	},
 }
 
 // Execute adds all child commands to the root command sets flags appropriately.
@@ -149,28 +121,4 @@ func init() {
 	if err != nil {
 		util.Failed("Failed to read global config file %s: %v", globalconfig.GetGlobalConfigPath(), err)
 	}
-}
-
-func sentryNotSetupWarning() {
-	if version.SentryDSN == "" && globalconfig.DdevGlobalConfig.InstrumentationOptIn {
-		output.UserOut.Warning("Instrumentation is opted in, but SentryDSN is not available.")
-	}
-}
-
-// checkDdevVersionAndOptInSentry() reads global config and checks to see if current version is different
-// from the last saved version. If it is, prompt to request anon ddev usage stats
-// and update the info.
-func checkDdevVersionAndOptInSentry() error {
-	if !output.JSONOutput && version.COMMIT != globalconfig.DdevGlobalConfig.LastUsedVersion && globalconfig.DdevGlobalConfig.InstrumentationOptIn == false && !globalconfig.DdevNoSentry {
-		allowStats := util.Confirm("It looks like you have a new ddev release.\nMay we send anonymous ddev usage statistics and errors?\nTo know what we will see please take a look at\nhttps://ddev.readthedocs.io/en/latest/users/cli-usage/#opt-in-usage-information\nPermission to beam up?")
-		if allowStats {
-			globalconfig.DdevGlobalConfig.InstrumentationOptIn = true
-		}
-		globalconfig.DdevGlobalConfig.LastUsedVersion = version.VERSION
-		err := globalconfig.WriteGlobalConfig(globalconfig.DdevGlobalConfig)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
