+++ pkg/output/text_formatter.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"github.com/drud/ddev/pkg/globalconfig"
 	"github.com/drud/ddev/pkg/version"
-	"github.com/getsentry/raven-go"
 	"io"
 	"os"
 	"sort"
@@ -128,7 +127,7 @@ func (f *TextFormatter) Format(entry *log.Entry) ([]byte, error) {
 	}
 
 	if entry.Level == log.FatalLevel && globalconfig.DdevGlobalConfig.InstrumentationOptIn && version.SentryDSN != "" {
-		_ = raven.CaptureMessageAndWait("Failed:"+entry.Message, map[string]string{"severity-level": "fatal", "report-type": "util-fail"})
+		//_ = raven.CaptureMessageAndWait("Failed:"+entry.Message, map[string]string{"severity-level": "fatal", "report-type": "util-fail"})
 	}
 	b.WriteByte('\n')
 	return b.Bytes(), nil
