+++ pkg/version/version.go
@@ -24,6 +24,10 @@ var DdevVersion = "v0.0.0-overridden-by-make" // Note that this is overridden by
 // It is compiled in using link-time variables
 var SentryDSN = ""
 
+// SegmentKey is the ddev-specific key for Segment service
+// Compiled with link-time variables
+var SegmentKey = ""
+
 // DockerVersionConstraint is the current minimum version of docker required for ddev.
 // See https://godoc.org/github.com/Masterminds/semver#hdr-Checking_Version_Constraints
 // for examples defining version constraints.
